% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addBroadPeakColumnNames.R
\name{add_broadPeak_columnNames}
\alias{add_broadPeak_columnNames}
\title{add broadPeak file column names to a GRanges object}
\usage{
add_broadPeak_columnNames(x)
}
\arguments{
\item{x}{GRanges object}
}
\value{
GRanges object
}
\description{
The chipSeeker readPeakFile function creates a GRanges object but the narrowPeak column names are missing.
This function adds them in.
}
