% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExpValueAndSample.R
\name{addExpValueAndSample}
\alias{addExpValueAndSample}
\title{addExpValueAndSample}
\usage{
addExpValueAndSample(x, y, z)
}
\arguments{
\item{x}{output from addGeneIdToIndex()}

\item{y}{output from splitIntoBins(makeAx4TPM())}

\item{z}{output from makeAx4TPM()}
}
\value{
Modified form of addGeneIdToIndex() output where mean and sample columns are added.
}
\description{
Adds mean expression value and sample based on expression level.
Internally calls lApplyFunction(), bin_search() & addSampleToNewList()
}
