% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addExtraAnnotation.R
\name{addExtraAnnotation}
\alias{addExtraAnnotation}
\title{addExtraAnnotation}
\usage{
addExtraAnnotation(x)
}
\arguments{
\item{x}{csAnno object which is the output of annotatePeaksSeq}
}
\value{
csAnno object
}
\description{
This function looks at the annotation column which is the output of the ChIPseeker::annotatePeak functon.
I've wrapped the ChIPseeker::annotatePeak function in annotatePeaksSeq.
The annotation can be a number of things promoter, exon 1, exon 2 etc. downstream 1kb etc.
}
\details{
The ChIPseeker::annotatePeak assigns a peak to a gene hierachically. First to a promoter, then exon, then downstream.
So if the annotation of a peak is exon 1, 2 or 3 of a gene, then the geneID present in the annotation column for said
peak should match that of the DDB_G0 in the geneID column.

This function asks this question. It extracts the geneID from the annotation column using regex and the puts it in
a separate column. It then creates a second new column 'annotationGeneIDEqualGeneID' that tells us if the two geneIDs match.
If they do match the value is "Same geneID" if they don't "Different annotation".

Other annotations such as promoter or downstream don't have a geneID in the original annotation column and so these
are given the value "Different annotation" in the 'annotationGeneIDEqualGeneID' column
}
