% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGeneIdToIndex.R
\name{addGeneIdToIndex}
\alias{addGeneIdToIndex}
\title{addGeneIdToIndex}
\usage{
addGeneIdToIndex(x, y)
}
\arguments{
\item{x}{output from selectPromoterGRangesBasedOnTagMatrixIndex()}

\item{y}{output from makeDictyGenesGRangesObject()}
}
\value{
List. Each List also a list containing two GRanges. The first GRange is the interesting part. Contains the Windows / Promoters GRange from
selectPromoterGRangesBasedOnTagMatrixIndex() but with GeneId added. You can chgeck against TestGRanges I think they should be the same.
}
\description{
Shrinks Genes GRange down to 3bp width around the TSS. The Genes GRAnge has a geneID associated with it.
Individual promoter GRanges are shrunk to 3bp around the TSS as well.
In theory the promoter and gene GRanges are the same. E.g.  DDB0232428 100-103 is the same in both dataframes.
We can now connect the index in the promoter GRange to a geneID!
}
