% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaksSeq.R
\name{annotatePeaksSeq}
\alias{annotatePeaksSeq}
\title{annotatePeaksSeq}
\usage{
annotatePeaksSeq(x, tssRegionStart = -3000, tssRegionEnd = 3000)
}
\arguments{
\item{x}{GRanges object}

\item{tssRegionStart}{number : default is -3000bp - what is set by ChIPseeker::annotatePeak - can be changed!}

\item{tssRegionEnd}{number : default is 3000bp - what is set by ChIPseeker::annotatePeak - can be changed!}
}
\value{

}
\description{
Takes in GRanges object of ChIP-seq peaks and annotates them based on a D.Discoideum TxDb object
coming from makeDictyGrangesfromTxDb.
}
\details{
The seqnames returned by makeDictyGrangesfromTxDb are 1, 2, 3, 4, 5, 6 and so I change the input seqnames
to match before converting them back to what they were previously after.
}
