% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropDictyChromosomes.R
\name{dropDictyChromosomes}
\alias{dropDictyChromosomes}
\title{dropDictyChromsomes}
\usage{
dropDictyChromosomes(x)
}
\arguments{
\item{x}{Granges object}
}
\value{
Granges object
}
\description{
Function to drop any chromsomes which are not d. discoideum chromsomes 1 through 6
Extra chromosomes and contigs - ribosomal, mitochondrial etc. may be interesinting but this
is just for simplicity
}
