% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/willAmendedGetTagMatrix.R
\name{getTagMatrixWill}
\alias{getTagMatrixWill}
\title{getTagMatrixWill}
\usage{
getTagMatrixWill(peak, weightCol = NULL, windows, flip_minor_strand = TRUE)
}
\arguments{
\item{peak}{peak GRanges file. Will probably be in the anno GRange from the output of annotatePeaksSeq or addExtraAnnotation}

\item{weightCol}{?}

\item{windows}{GRanges object of promoters - output from makeDictyPromotersGRangesObject()}

\item{flip_minor_strand}{flips tags so there all in the same direction. E.g if the output was 000000111111 and the strand is negative it will become 111111000000}
}
\value{
list of a tagMatrix, which will be used to make a plot and a GRanges object of windows / promoters.
This GRanges object now has the index column. The index column should map between the tagMatrix rowname and the value in this column.
This can be used to reverse engineer what promoter belongs to what gene.
}
\description{
getTagMatrixWill
}
