% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_pipeline.R
\name{mainPipeLine}
\alias{mainPipeLine}
\title{mainPipeLine}
\usage{
mainPipeLine(x, y, bins = 10)
}
\arguments{
\item{x}{filepath to narrowPeak or broadPeak file}

\item{y}{filepath to 6 column un-normalised ax4 set1 readcounts file}

\item{bins}{number of bins}
}
\value{

}
\description{
mainPipeLine
}
