% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDictyGenesGRangesObject.R
\name{makeDictyGenes}
\alias{makeDictyGenes}
\title{makeDictyGenesGRangesObject}
\usage{
makeDictyGenes()
}
\value{
A list of GRanges. First is a GRanges of all dicty genes.
Then just positive strand dicty genes and then negative strand dicty genes.
}
\description{
Calls makeDictyGRangesfromTxdb(). The gets the genes GRanges from that.
}
