% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeDictyPromotersGRangesObject.R
\name{makeDictyPromotersGrangesObject}
\alias{makeDictyPromotersGrangesObject}
\title{makeDictyPromotersGrangesObject}
\usage{
makeDictyPromotersGrangesObject(Start = 2500, End = 2500)
}
\arguments{
\item{Start}{defaults to 2500 upstream}

\item{End}{defaults to 2500 downstream}
}
\value{
GRanges Object of D. Discoideum promoters for chromsomes 1 to 6
}
\description{
Can change what is defined as a promoter by altering the Start and End parameters. I've basically wrapped the ChIPseeker::getPromoters function and then removed chromosomses 1 to 6.
}
