% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctionWillDictyChip.R
\name{plotFunctionWillDictyChip}
\alias{plotFunctionWillDictyChip}
\title{plotFunctionWillDictyChip}
\usage{
plotFunctionWillDictyChip(x)
}
\arguments{
\item{x}{Takes in the output of rearrangeTagMatricesBasedOnSample() which is a list of tagN}
}
\value{
Dataframe
}
\description{
plotFunctionWillDictyChip
}
