% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rearrangeTagMatricesBasedOnSample.R
\name{rearrangeTagMatricesBasedOnSample}
\alias{rearrangeTagMatricesBasedOnSample}
\title{rearrangeTagMatricesBasedOnSample}
\usage{
rearrangeTagMatricesBasedOnSample(x, y)
}
\arguments{
\item{x}{output of addExpValueAndSample()}

\item{y}{output of getTagMatricesForDiffExpLevelBins()

This is a list of GRanges objects based on the tagMatrices from getTagMatricesForDiffExpLevelBins().They have the index, geneId, mean, sample as metadata columns.
The sample metadata column is based expression level}
}
\value{
output is the correct tagMatrices! As in matrices where
}
\description{
rearrangeTagMatricesBasedOnSample
}
