% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectPromoterGRangesBasedOnTagMatrixIndex.R
\name{selectPromoterGRangesBasedOnTagMatrixIndex}
\alias{selectPromoterGRangesBasedOnTagMatrixIndex}
\title{selectPromoterGRangesBasedOnTagMatrixIndex}
\usage{
selectPromoterGRangesBasedOnTagMatrixIndex(x)
}
\arguments{
\item{x}{Output from getTagMatricesForDiffExpLevelBins which will likely be a list of tagMatrices and Windows / Promoters GRanges with index column}
}
\value{
A new list. Each element is based on one of the incoming tagMatrices in the input list. Each element in the new list is also a list. These elements show
the Promoter GRange object subsetted by the rownames of the tagMatrix. Where the rowname of the tagMatrix matches the index column in the Windows / Promoter GRange
object, that row is returned. This is also done in a strand specific way to generate + and - strand Promoter GRange objects.
}
\description{
selectPromoterGRangesBasedOnTagMatrixIndex
}
