% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set1DegOverlapNumbers.R
\name{set1DegOverlapNumbers}
\alias{set1DegOverlapNumbers}
\title{Title}
\usage{
set1DegOverlapNumbers(x, set1_up, set1_down)
}
\arguments{
\item{x}{List of rearranged GRange objects. Output from The sample column should all be the same e.g. 1 / 2 ... 9. If its a mixture this isn't the right list}

\item{set1_up}{dataframe}

\item{set1_down}{dataframe}
}
\value{

}
\description{
Title
}
