% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitIntoBins.R
\name{SplitIntoBins}
\alias{SplitIntoBins}
\title{SplitIntoBins}
\usage{
SplitIntoBins(x, z)
}
\arguments{
\item{x}{data frame to be split}

\item{z}{number - how many bins you wish to split into}
}
\value{

}
\description{
SplitIntoBins
}
